requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
docker-pycreds==0.4.0
easyocr==1.7.2
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filelock==3.17.0
fire==0.7.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.1
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==2.8.8
ninja==1.11.1.3
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.10.0
paddlepaddle==2.6.2
pandas==2.2.3
pathtools==0.1.2
pillow==11.1.0
platformdirs==4.3.6
promise==2.3
protobuf==3.20.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.2
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.22.0
setproctitle==1.3.5
shapely==2.0.7
shortuuid==1.0.13
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stringzilla==3.12.2
sympy==1.13.3
termcolor==2.3.0
tifffile==2024.8.30
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
-e git+ssh://git@github.com/ultralytics/ultralytics.git@23a90142dc66fbb180fe1bb513a2adc44322c978#egg=ultralytics
ultralytics-thop==2.0.14
urllib3==2.3.0
wandb==0.19.8
yaspin==3.1.0
zipp==3.21.0
